fix incorrect schedule parsing
Now lessons text is grouped before beign added to the schedule
so overlaping text doesn't end up in a different lesson.
Also fixed incorrect bounds calculation for the name of the group when
it is located near the other column.
And big lesson names are automatically expanded to fit maximum possible width,
since this is the most common pattern for them, but it's not detectible
by looking at the text alone